refactoring rename AAA to PhonesContext
